pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.73.1",
    "google-generativeai>=0.8.6",
    "llama-parse>=0.6.94",
    "llmwhisperer-client>=2.3.1",
    "pandas>=3.0.2",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.82.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]
